requirements.txt
Django>=4.2
djangorestframework>=3.15.1
locust>=2.24.1
Faker>=24.8.0
djangorestframework-simplejwt>=5.3.1